Moved the `ACCEPTED_INFECTION_LABEL`/`FORECASTED_INFECTION_LABEL` constants to the external `computed_statistics` module, and replaced their string literal uses with the corresponding constant. Added additional doc comments.
