+++ src/computed_statistics.rs
@@ -43,9 +43,22 @@
 
 use crate::profiling::{
     add_computed_statistic, CustomStatisticComputer, CustomStatisticPrinter, ProfilingData,
-    ACCEPTED_INFECTION_LABEL, FORECASTED_INFECTION_LABEL,
 };
 
+/// The name of the distinguished accepted infection label. You don't need to make a constant
+/// for this, but it can keep you from introducing a bug because of a typo in the label.
+/// You would use the constant everywhere in your code instead of the string literal.
+pub const ACCEPTED_INFECTION_LABEL: &str = "accepted infection attempt";
+/// The name of the distinguished forecasted infection label.
+pub const FORECASTED_INFECTION_LABEL: &str = "forecasted infection";
+
+/// The function that knows how to compute the forecasting efficiency from the
+/// data collected in `ProfilingData`.
+///
+/// - This could just as easily be a closure, but making it a function can make
+///   things a lot more readable.
+/// - This function returns an option to accommodate statistics that are only
+///   conditionally computable.
 fn forecasting_efficiency_computer(statistics: &ProfilingData) -> Option<f64> {
     if let (Some(&accepted), Some(&forecasted)) = (
         statistics.counts.get(&ACCEPTED_INFECTION_LABEL),
@@ -59,6 +72,9 @@ fn forecasting_efficiency_computer(statistics: &ProfilingData) -> Option<f64> {
     }
 }
 
+/// The function that knows how to display the forecasting efficiency on the console.
+/// This could just as easily be a closure if you prefer. The type inside the option
+/// returned by the computer and argument of the printer function must coincide.
 fn forecasting_efficiency_printer(efficiency: f64) {
     println!("Infection Forecasting Efficiency: {:.2}%", efficiency);
 }
@@ -77,4 +93,22 @@ pub fn init() {
 
     // Use the free function in the `profiling` module to register the statistic.
     add_computed_statistic(label, description, computer, printer);
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::profiling::{get_profiling_data, print_named_counts};
+    use std::time::{Duration, Instant};
+
+    #[test]
+    fn print_named_counts_computes_forecast_efficiency() {
+        {
+            let mut data = get_profiling_data();
+            data.start_time = Some(Instant::now().checked_sub(Duration::from_secs(2)).unwrap());
+            data.counts.insert(FORECASTED_INFECTION_LABEL, 10);
+            data.counts.insert(ACCEPTED_INFECTION_LABEL, 4);
+        }
+        print_named_counts(); // should print "40.00% efficiency"
+    }
 }
